refactor: optimized code
