+++ src/assets/css/github-code.css
@@ -1,11 +1,11 @@
 pre code.hljs {
-    display: block;
-    overflow-x: auto;
-    padding: 1em
+  display: block;
+  overflow-x: auto;
+  padding: 1em;
 }
 
 code.hljs {
-    padding: 3px 5px
+  padding: 3px 5px;
 }
 
 /*!
@@ -19,83 +19,109 @@ code.hljs {
   Current colors taken from GitHub's CSS
 */
 .hljs {
-    color: #24292e;
-    background: transparent;
+  color: #24292e;
+  background: transparent;
 }
 
-.hljs-doctag, .hljs-keyword, .hljs-meta .hljs-keyword, .hljs-template-tag, .hljs-template-variable, .hljs-type, .hljs-variable.language_ {
-    color: #d73a49
+.hljs-doctag,
+.hljs-keyword,
+.hljs-meta .hljs-keyword,
+.hljs-template-tag,
+.hljs-template-variable,
+.hljs-type,
+.hljs-variable.language_ {
+  color: #d73a49;
 }
 
-.hljs-title, .hljs-title.class_, .hljs-title.class_.inherited__, .hljs-title.function_ {
-    color: #6f42c1
+.hljs-title,
+.hljs-title.class_,
+.hljs-title.class_.inherited__,
+.hljs-title.function_ {
+  color: #6f42c1;
 }
 
-.hljs-attr, .hljs-attribute, .hljs-literal, .hljs-meta, .hljs-number, .hljs-operator, .hljs-selector-attr, .hljs-selector-class, .hljs-selector-id, .hljs-variable {
-    color: #005cc5
+.hljs-attr,
+.hljs-attribute,
+.hljs-literal,
+.hljs-meta,
+.hljs-number,
+.hljs-operator,
+.hljs-selector-attr,
+.hljs-selector-class,
+.hljs-selector-id,
+.hljs-variable {
+  color: #005cc5;
 }
 
-.hljs-meta .hljs-string, .hljs-regexp, .hljs-string {
-    color: #032f62
+.hljs-meta .hljs-string,
+.hljs-regexp,
+.hljs-string {
+  color: #032f62;
 }
 
-.hljs-built_in, .hljs-symbol {
-    color: #e36209
+.hljs-built_in,
+.hljs-symbol {
+  color: #e36209;
 }
 
-.hljs-code, .hljs-comment, .hljs-formula {
-    color: #6a737d
+.hljs-code,
+.hljs-comment,
+.hljs-formula {
+  color: #6a737d;
 }
 
-.hljs-name, .hljs-quote, .hljs-selector-pseudo, .hljs-selector-tag {
-    color: #22863a
+.hljs-name,
+.hljs-quote,
+.hljs-selector-pseudo,
+.hljs-selector-tag {
+  color: #22863a;
 }
 
 .hljs-subst {
-    color: #24292e
+  color: #24292e;
 }
 
 .hljs-section {
-    color: #005cc5;
-    font-weight: 700
+  color: #005cc5;
+  font-weight: 700;
 }
 
 .hljs-bullet {
-    color: #735c0f
+  color: #735c0f;
 }
 
 .hljs-emphasis {
-    color: #24292e;
-    font-style: italic
+  color: #24292e;
+  font-style: italic;
 }
 
 .hljs-strong {
-    color: #24292e;
-    font-weight: 700
+  color: #24292e;
+  font-weight: 700;
 }
 
 .hljs-addition {
-    color: #22863a;
-    background-color: #f0fff4
+  color: #22863a;
+  background-color: #f0fff4;
 }
 
 .hljs-deletion {
-    color: #b31d28;
-    background-color: #ffeef0
+  color: #b31d28;
+  background-color: #ffeef0;
 }
 
 .dark {
-    pre code.hljs {
-        display: block;
-        overflow-x: auto;
-        padding: 1em
-    }
+  pre code.hljs {
+    display: block;
+    overflow-x: auto;
+    padding: 1em;
+  }
 
-    code.hljs {
-        padding: 3px 5px
-    }
+  code.hljs {
+    padding: 3px 5px;
+  }
 
-    /*!
+  /*!
       Theme: GitHub Dark
       Description: Dark theme as seen on github.com
       Author: github.com
@@ -106,70 +132,95 @@ code.hljs {
       Current colors taken from GitHub's CSS
     */
 
-    .hljs {
-        color: #c9d1d9;
-        background: transparent;
-    }
+  .hljs {
+    color: #c9d1d9;
+    background: transparent;
+  }
 
-    .hljs-doctag, .hljs-keyword, .hljs-meta .hljs-keyword, .hljs-template-tag, .hljs-template-variable, .hljs-type, .hljs-variable.language_ {
-        color: #ff7b72
-    }
+  .hljs-doctag,
+  .hljs-keyword,
+  .hljs-meta .hljs-keyword,
+  .hljs-template-tag,
+  .hljs-template-variable,
+  .hljs-type,
+  .hljs-variable.language_ {
+    color: #ff7b72;
+  }
 
-    .hljs-title, .hljs-title.class_, .hljs-title.class_.inherited__, .hljs-title.function_ {
-        color: #d2a8ff
-    }
+  .hljs-title,
+  .hljs-title.class_,
+  .hljs-title.class_.inherited__,
+  .hljs-title.function_ {
+    color: #d2a8ff;
+  }
 
-    .hljs-attr, .hljs-attribute, .hljs-literal, .hljs-meta, .hljs-number, .hljs-operator, .hljs-selector-attr, .hljs-selector-class, .hljs-selector-id, .hljs-variable {
-        color: #79c0ff
-    }
+  .hljs-attr,
+  .hljs-attribute,
+  .hljs-literal,
+  .hljs-meta,
+  .hljs-number,
+  .hljs-operator,
+  .hljs-selector-attr,
+  .hljs-selector-class,
+  .hljs-selector-id,
+  .hljs-variable {
+    color: #79c0ff;
+  }
 
-    .hljs-meta .hljs-string, .hljs-regexp, .hljs-string {
-        color: #a5d6ff
-    }
+  .hljs-meta .hljs-string,
+  .hljs-regexp,
+  .hljs-string {
+    color: #a5d6ff;
+  }
 
-    .hljs-built_in, .hljs-symbol {
-        color: #ffa657
-    }
+  .hljs-built_in,
+  .hljs-symbol {
+    color: #ffa657;
+  }
 
-    .hljs-code, .hljs-comment, .hljs-formula {
-        color: #8b949e
-    }
+  .hljs-code,
+  .hljs-comment,
+  .hljs-formula {
+    color: #8b949e;
+  }
 
-    .hljs-name, .hljs-quote, .hljs-selector-pseudo, .hljs-selector-tag {
-        color: #7ee787
-    }
+  .hljs-name,
+  .hljs-quote,
+  .hljs-selector-pseudo,
+  .hljs-selector-tag {
+    color: #7ee787;
+  }
 
-    .hljs-subst {
-        color: #c9d1d9
-    }
+  .hljs-subst {
+    color: #c9d1d9;
+  }
 
-    .hljs-section {
-        color: #1f6feb;
-        font-weight: 700
-    }
+  .hljs-section {
+    color: #1f6feb;
+    font-weight: 700;
+  }
 
-    .hljs-bullet {
-        color: #f2cc60
-    }
+  .hljs-bullet {
+    color: #f2cc60;
+  }
 
-    .hljs-emphasis {
-        color: #c9d1d9;
-        font-style: italic
-    }
+  .hljs-emphasis {
+    color: #c9d1d9;
+    font-style: italic;
+  }
 
-    .hljs-strong {
-        color: #c9d1d9;
-        font-weight: 700
-    }
+  .hljs-strong {
+    color: #c9d1d9;
+    font-weight: 700;
+  }
 
-    .hljs-addition {
-        color: #aff5b4;
-        background-color: #033a16
-    }
+  .hljs-addition {
+    color: #aff5b4;
+    background-color: #033a16;
+  }
 
-    .hljs-deletion {
-        color: #ffdcd7;
-        background-color: #67060c
-    }
+  .hljs-deletion {
+    color: #ffdcd7;
+    background-color: #67060c;
+  }
 }
-
