+++ src/assets/css/root.css
@@ -2,84 +2,82 @@
 @import url(input.css);
 
 @font-face {
-    font-family: 'Barlow';
-    src: url('/src/assets/fonts/Barlow-Regular.otf');
-    font-style: normal;
-    font-display: fallback;
+  font-family: 'Barlow';
+  src: url('/src/assets/fonts/Barlow-Regular.otf');
+  font-style: normal;
+  font-display: fallback;
 }
 
 @font-face {
-    font-family: 'Barlow';
-    src: url('/src/assets/fonts/Barlow-Medium.otf');
-    font-weight: 500;
-    font-style: normal;
-    font-display: fallback;
+  font-family: 'Barlow';
+  src: url('/src/assets/fonts/Barlow-Medium.otf');
+  font-weight: 500;
+  font-style: normal;
+  font-display: fallback;
 }
 
 @font-face {
-    font-family: 'Barlow';
-    src: url('/src/assets/fonts/Barlow-Bold.otf');
-    font-weight: 700;
-    font-style: normal;
-    font-display: fallback;
+  font-family: 'Barlow';
+  src: url('/src/assets/fonts/Barlow-Bold.otf');
+  font-weight: 700;
+  font-style: normal;
+  font-display: fallback;
 }
 
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
 
-
 @layer base {
-    body {
-        font-family: 'Barlow', ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI',
-        Roboto, 'Helvetica Neue', Arial, 'Noto Sans', sans-serif, 'Apple Color Emoji',
-        'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
-        @apply bg-light-background dark:bg-dark-background;
-        @apply text-light-secondary dark:text-dark-secondary text-sm;
+  body {
+    font-family: 'Barlow', ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI',
+      Roboto, 'Helvetica Neue', Arial, 'Noto Sans', sans-serif, 'Apple Color Emoji',
+      'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
+    @apply bg-light-background dark:bg-dark-background;
+    @apply text-light-secondary dark:text-dark-secondary text-sm;
+  }
+
+  .link {
+    @apply cursor-pointer transition duration-150 ease-in-out;
+
+    &:hover {
+      @apply opacity-70;
     }
+  }
 
-    .link {
-        @apply cursor-pointer transition duration-150 ease-in-out;
+  a {
+    @apply link text-primary;
+  }
 
-        &:hover {
-            @apply opacity-70;
-        }
-    }
+  :focus-visible {
+    outline: none;
+  }
 
-    a {
-        @apply link text-primary;
-    }
+  .bg-gradient-primary {
+    @apply bg-gradient-to-l from-success to-primary;
+  }
 
-    :focus-visible {
-        outline: none;
-    }
+  .bg-gradient-primary-t {
+    @apply from-success to-primary bg-gradient-to-t;
+  }
 
-    .bg-gradient-primary {
-        @apply bg-gradient-to-l from-success to-primary;
-    }
+  .bg-gradient-primary-tl {
+    @apply from-success to-primary bg-gradient-to-tl;
+  }
 
-    .bg-gradient-primary-t {
-        @apply from-success to-primary bg-gradient-to-t;
-    }
+  .bg-gradient-primary-tr {
+    @apply from-success to-primary bg-gradient-to-tr;
+  }
 
-    .bg-gradient-primary-tl {
-        @apply from-success to-primary bg-gradient-to-tl;
-    }
+  .bg-gradient-primary-b {
+    @apply from-success to-primary bg-gradient-to-b;
+  }
 
-    .bg-gradient-primary-tr {
-        @apply from-success to-primary bg-gradient-to-tr;
-    }
+  .bg-gradient-primary-bl {
+    @apply from-success to-primary bg-gradient-to-bl;
+  }
 
-    .bg-gradient-primary-b {
-        @apply from-success to-primary bg-gradient-to-b;
-    }
-
-    .bg-gradient-primary-bl {
-        @apply from-success to-primary bg-gradient-to-bl;
-    }
-
-    .bg-gradient-primary-br {
-        @apply from-success to-primary bg-gradient-to-br;
-    }
-
+  .bg-gradient-primary-br {
+    @apply from-success to-primary bg-gradient-to-br;
+  }
 }
