+++ src/common/constants/site-basic.ts
@@ -1,5 +1,5 @@
 export const SITE_GITHUB_URL = 'https://github.com/Flouse/ckb-school-site';
-export const ZERO2_URL = 'https://zero2ckb.ckbapp.dev/learn';
+export const ZERO2CKB_URL = 'https://zero2ckb.ckbapp.dev/learn';
 
 export enum LangsEnum {
   en = 'en',
