+++ src/components/Loading.tsx
@@ -1,12 +1,12 @@
 import { Component, JSX, ParentProps } from 'solid-js';
 import { FaSolidSpinner } from 'solid-icons/fa';
 
-interface ILoadingProps extends ParentProps {
+interface Props extends ParentProps {
   title?: JSX.Element;
   icon?: JSX.Element;
 }
 
-const Loading: Component<ILoadingProps> = (props) => {
+const Loading: Component<Props> = (props) => {
   return (
     <div class="flex flex-col items-center">
       {props.icon || <FaSolidSpinner class="animate-spin text-7xl mb-6" />}
