+++ src/components/Toast/ToastContext.tsx
@@ -11,10 +11,10 @@ import * as zagToast from '@zag-js/toast';
 import { normalizeProps, useMachine } from '@zag-js/solid';
 import Toast from '~/components/Toast/index';
 import { Options, Toaster } from '@zag-js/toast/dist/toast.types';
-import { IToastContext } from '~/components/Toast/types';
+import { ToastContext } from '~/components/Toast/types';
 
-const ToastContext = createContext<IToastContext>();
-export const useToast = () => useContext(ToastContext)!;
+const toastContext = createContext<ToastContext>();
+export const useToast = () => useContext(toastContext)!;
 export const toast: Toaster = zagToast.api()!;
 
 const ToastProvider: ParentComponent = (props) => {
@@ -22,7 +22,7 @@ const ToastProvider: ParentComponent = (props) => {
     zagToast.group.machine({ offsets: '10px', id: createUniqueId() }),
   );
   const api = createMemo(() => zagToast.group.connect(state, send, normalizeProps));
-  const context: IToastContext = {
+  const context: ToastContext = {
     ...api(),
     create: (options: Options) => {
       return api().create({
@@ -46,7 +46,7 @@ const ToastProvider: ParentComponent = (props) => {
   };
 
   return (
-    <ToastContext.Provider value={context}>
+    <toastContext.Provider value={context}>
       <For each={Object.entries(api().toastsByPlacement)}>
         {([placement, toasts]) => {
           return (
@@ -57,7 +57,7 @@ const ToastProvider: ParentComponent = (props) => {
         }}
       </For>
       {props.children}
-    </ToastContext.Provider>
+    </toastContext.Provider>
   );
 };
