+++ src/components/Toast/index.css
@@ -1,8 +1,8 @@
 @keyframes progress {
-    0% {
-        width: 100%;
-    }
-    100% {
-        width: 0;
-    }
+  0% {
+    width: 100%;
+  }
+  100% {
+    width: 0;
+  }
 }
