+++ src/content/courses/basic-theory/chapter_1.mdx
@@ -8,9 +8,8 @@ Unspent cells are live cells; spent cells are dead cells. So, a `CKB` chain keep
 
 ![Figure 1. Layered Architecture](https://images.pexels.com/photos/11198507/pexels-photo-11198507.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=2)
 
->In the CKB universe, countless cells are floating around.
-Cells store data that jointly form the general state of the chain.
-
+> In the CKB universe, countless cells are floating around.
+> Cells store data that jointly form the general state of the chain.
 
 Unlike `UTXO`, a cell can store any type of data. Each cell has a field called data, where you can put an
 unformatted string. The string can be in whatever format you want, as long as you can interpret it.
