+++ src/content/courses/basic-theory/chapter_3.mdx
@@ -12,14 +12,14 @@ In essence, the scripts are a piece of code and parameters. When we try to consu
 
 The script structure looks like this:
 
-
-``` ts
+```ts
 Script: {
   code_hash: HexString
   args: HexString
   hash_type: either 'data' or 'type'
 }
 ```
+
 In these three fields, hash_type is reserved for later, the other two are:
 
 - `code_hash`: the hash of a certain piece of code
