+++ src/content/courses/basic-theory/chapter_5.mdx
@@ -4,7 +4,7 @@ We have learned that the lock and type fields of the cell are used to lock the b
 
 Lock is a script structure that looks like this:
 
-``` ts
+```ts
 Script: {
   code_hash: HexString
   args: HexString
