+++ src/content/courses/basic-theory/chapter_8.mdx
@@ -14,4 +14,3 @@ This variance in use comes down to the difference in the design of the two locks
 - `Type script` In a transaction, the type scripts of all inputs and outputs will be executed once.
 
 Due to the variations in execution mechanisms, derived usage also varies. You are free to have your own opinions on this, but essentially these are the recommended official usages.
-
