+++ src/content/courses/community-courses.ts
@@ -1,0 +1,3 @@
+import { Course } from '~/types/course';
+
+export const communityCourses: Course[] = [];
