+++ src/pages/CourseDetails.tsx
@@ -1,13 +1,13 @@
 import { Component, ParentProps, Show } from 'solid-js';
 import { useRouteData } from '@solidjs/router';
-import { ICourseDetailsData } from '~/pages/CourseDetails.data';
-import CourseExplorer from '~/components/Course/CourseExplorer';
-import { CourseProvider } from '~/components/Course/CourseContext';
+import { CourseDetailsData } from '~/pages/CourseDetails.data';
+import Index from '~/components/CourseCore/CourseExplorer';
+import { CourseProvider } from '~/components/CourseCore/CourseContext';
 import Loading from '~/components/Loading';
 import DataEmpty from '~/components/DataEmpty';
 
 const CourseDetails: Component<ParentProps> = () => {
-  const { course, loading } = useRouteData<ICourseDetailsData>();
+  const { course, loading } = useRouteData<CourseDetailsData>();
   return (
     <>
       <Show
@@ -29,7 +29,7 @@ const CourseDetails: Component<ParentProps> = () => {
           }
         >
           <CourseProvider course={course!}>
-            <CourseExplorer />
+            <Index />
           </CourseProvider>
         </Show>
       </Show>
