+++ src/pages/Playground.tsx
@@ -1,7 +1,4 @@
 import { Component, ParentProps } from 'solid-js';
-import { Motion } from '@motionone/solid';
-import { A } from 'solid-start';
-import { BiLogosGithub } from 'solid-icons/bi';
 import EmptyPage from '~/components/EmptyPage';
 
 const Playground: Component<ParentProps> = () => {
