--- src/types/interfaces.ts
@@ -1,5 +1,0 @@
-export interface IBaseLink {
-  title: string;
-  to: string;
-  target?: '_blank' | '_top' | '_parent' | '_self';
-}
