+++ src/utils/ckb-provider.ts
@@ -1,8 +1,8 @@
-export interface ICKBProvider {
+export interface CKBProvider {
   send(): void;
 }
 
-export default class CkbProvider implements ICKBProvider {
+export default class CkbProvider implements CKBProvider {
   send() {
     console.log('send');
   }
